refactor(node): extract canvas interactive routes into canvas-interactive.ts (task-1773689755389)
- 6 routes extracted: gaze, briefing, victory, spark, express, render/stream
- Reality Mixer infrastructure moved (broadcastRenderCommand, renderCommandLog, subscribers)
- Auto-expression event listener moved to plugin
- server.ts reduced by 446 lines
- Total extracted routes: 19 (canvas-routes: 13, canvas-takeover: 3, canvas-interactive: 6, minus 3 overlap)
- Route-docs contract: 575/575 pass
- Tests: 2455 pass
