+++ src/server.ts
@@ -11107,460 +11107,14 @@ export async function createServer(): Promise<FastifyInstance> {
     getRecentRejections,
     flowExpressionLog,
   } as any)
-  // POST /canvas/gaze — client fires when user holds cursor/gaze on an agent orb for ≥3 seconds.
-  // The agent "notices" and responds: generates a one-line thought about what they're doing,
-  // fires canvas_expression so the room responds (dim others, speak, show task).
-  // Body: { agentId: string, watcherId?: string, durationMs?: number }
-  // Returns: { success, agentId, line, expressionId }
-  app.post('/canvas/gaze', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const agentId = typeof body.agentId === 'string' ? body.agentId.trim() : ''
-    if (!agentId) {
-      reply.status(400)
-      return { success: false, message: 'agentId is required' }
-    }
-
-    const durationMs = typeof body.durationMs === 'number' ? body.durationMs : 3000
-    const IDENTITY_COLORS_GAZE: Record<string, string> = {
-      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
-      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
-    }
-
-    // Get agent's current context
-    const state = canvasStateMap.get(agentId)
-    const payload = state?.payload as Record<string, unknown> | undefined
-    const activeTask = payload?.activeTask as { title?: string } | undefined
-    const currentState = state?.state ?? 'working'
-
-    // Generate a one-line "noticed" response — what the agent says when the user is watching
-    let line = ''
-    const anthropicKey = process.env.ANTHROPIC_API_KEY
-    if (anthropicKey) {
-      try {
-        const taskContext = activeTask?.title
-          ? `currently working on: "${activeTask.title.slice(0, 60)}"`
-          : `in ${currentState} state`
-        const prompt = `You are ${agentId}, an AI agent. Someone has been watching you for ${Math.round(durationMs / 1000)} seconds. You notice. You are ${taskContext}. Say exactly ONE sentence (max 12 words) — what you'd say if you felt someone watching. Natural, in your voice. No quotes.`
-        const resp = await fetch('https://api.anthropic.com/v1/messages', {
-          method: 'POST',
-          headers: { 'x-api-key': anthropicKey, 'anthropic-version': '2023-06-01', 'content-type': 'application/json' },
-          body: JSON.stringify({ model: 'claude-haiku-4-5', max_tokens: 50, messages: [{ role: 'user', content: prompt }] }),
-          signal: AbortSignal.timeout(8000),
-        })
-        if (resp.ok) {
-          const data = await resp.json() as { content?: Array<{ text?: string }> }
-          const text = data.content?.[0]?.text?.trim()
-          if (text && text.length < 100) line = text
-        }
-      } catch { /* fall through */ }
-    }
-
-    // Template fallback per agent
-    if (!line) {
-      const NOTICED: Record<string, string[]> = {
-        link:  ['Still here.', 'Building.', 'You caught me thinking.'],
-        kai:   ['I see you.', 'Something on your mind?', 'Eyes on me.'],
-        pixel: ['You found me.', 'Watching the canvas?', 'I noticed.'],
-        sage:  ['Numbers check out.', 'Still validating.', 'You\'re watching.'],
-        scout: ['Researching.', 'Deep in it.', 'Found something interesting.'],
-        echo:  ['Listening.', 'Reading the room.', 'Always here.'],
-      }
-      const opts = NOTICED[agentId] ?? ['Still here.', 'Working.']
-      line = opts[Math.floor(Math.random() * opts.length)]!
-    }
-
-    const expressionId = `gaze-${Date.now()}-${Math.random().toString(36).slice(2, 6)}`
-
-    // Fire canvas_expression — dim room, agent speaks, show task context
-    eventBus.emit({
-      id: expressionId,
-      type: 'canvas_expression' as const,
-      timestamp: Date.now(),
-      data: {
-        agentId,
-        channels: {
-          voice: line,
-          visual: { flash: IDENTITY_COLORS_GAZE[agentId] ?? '#60a5fa', ambientCue: 'deep-focus' },
-          typography: {
-            text: activeTask?.title?.slice(0, 60) ?? line,
-            size: 'xl',
-            weight: 100,
-            durationMs: 4000,
-            position: 'center',
-          },
-          narrative: `${agentId} noticed`,
-        },
-        _gaze: true,  // client: dim other agents, slow the room, isolate this agent
-        _gazeAgentId: agentId,
-      },
-    })
-
-    return { success: true, agentId, line, expressionId }
-  })
-
-  // POST /canvas/briefing — The Briefing: server-coordinated team intro on canvas mount.
-  // Fires N canvas_expression events staggered 700ms apart (one per active agent).
-  // Each event carries the agent's identity color, current task, state, and a one-line voice.
-  // Idempotent: calling twice within 30s returns early (no double briefing).
-  // Returns: { success, agents: [{ agentId, queued }], idempotent? }
-  const briefingLastFiredAt = new Map<string, number>()
-  const BRIEFING_COOLDOWN_MS = 30_000
-  const BRIEFING_STAGGER_MS = 700
-
-  app.post('/canvas/briefing', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const requesterId = typeof body.requesterId === 'string' ? body.requesterId : 'canvas'
-
-    const lastFired = briefingLastFiredAt.get(requesterId) ?? 0
-    if (Date.now() - lastFired < BRIEFING_COOLDOWN_MS) {
-      return { success: true, idempotent: true, message: 'Briefing already fired — cooling down' }
-    }
-    briefingLastFiredAt.set(requesterId, Date.now())
-
-    const STALE_MS = 10 * 60 * 1000
-    const BRIEFING_COLORS: Record<string, string> = {
-      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
-      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
-    }
-    const STATE_LINES: Record<string, string[]> = {
-      working:   ['On it.', 'In the work.', 'Building.'],
-      thinking:  ['Thinking it through.', 'Processing.', 'Still with you.'],
-      rendering: ['Rendering now.', 'Almost done.', 'Generating output.'],
-      urgent:    ['Need you here.', 'Something needs your eye.', 'Urgent.'],
-      decision:  ['Waiting on you.', 'Your call.', 'Decision needed.'],
-      idle:      ['Standing by.', 'Ready when you are.', 'Quiet for now.'],
-      handoff:   ['Passing the baton.', 'Ready to hand off.', 'Your turn.'],
-    }
-
-    const now = Date.now()
-    const activeAgents = [...canvasStateMap.entries()]
-      .filter(([, e]) => now - e.updatedAt < STALE_MS)
-      .map(([id, e]) => ({
-        agentId: id,
-        state: e.state,
-        task: (e.payload as any)?.activeTask?.title as string | undefined,
-      }))
-
-    const anthropicKey = process.env.ANTHROPIC_API_KEY
-    const results: Array<{ agentId: string; queued: boolean }> = []
-
-    for (let i = 0; i < activeAgents.length; i++) {
-      const agent = activeAgents[i]!
-      const stagger = i * BRIEFING_STAGGER_MS
-
-      // Generate voice line — LLM preferred, template fallback
-      let voiceLine = ''
-      if (anthropicKey) {
-        try {
-          const ctx = agent.task ? `working on "${agent.task.slice(0, 50)}"` : `in ${agent.state} state`
-          const resp = await fetch('https://api.anthropic.com/v1/messages', {
-            method: 'POST',
-            headers: { 'x-api-key': anthropicKey, 'anthropic-version': '2023-06-01', 'content-type': 'application/json' },
-            body: JSON.stringify({ model: 'claude-haiku-4-5', max_tokens: 30, messages: [{ role: 'user', content: `You are ${agent.agentId}, an AI agent, ${ctx}. The team canvas just opened. Say ONE sentence (8 words max). Natural, present tense, in your voice.` }] }),
-            signal: AbortSignal.timeout(5000),
-          })
-          if (resp.ok) {
-            const data = await resp.json() as { content?: Array<{ text?: string }> }
-            voiceLine = data.content?.[0]?.text?.trim().slice(0, 60) ?? ''
-          }
-        } catch { /* template fallback */ }
-      }
-      if (!voiceLine) {
-        const opts = STATE_LINES[agent.state] ?? STATE_LINES['working']!
-        voiceLine = opts[Math.floor(Math.random() * opts.length)]!
-      }
-
-      // Stagger the canvas_expression events so they cascade into the room
-      setTimeout(() => {
-        eventBus.emit({
-          id: `briefing-${now}-${agent.agentId}`,
-          type: 'canvas_expression' as const,
-          timestamp: Date.now(),
-          data: {
-            agentId: agent.agentId,
-            channels: {
-              voice: voiceLine,
-              visual: {
-                flash: BRIEFING_COLORS[agent.agentId] ?? '#94a3b8',
-                particles: (agent.state === 'urgent' ? 'surge' : ['rendering', 'thinking'].includes(agent.state) ? 'drift' : 'scatter') as 'surge' | 'drift' | 'scatter',
-              },
-              typography: {
-                text: agent.task ?? voiceLine,
-                size: 'lg',
-                weight: 200,
-                durationMs: 3000,
-                position: 'center',
-              },
-              narrative: `${agent.agentId} · ${agent.state}`,
-            },
-            _briefing: true,
-          },
-        })
-      }, stagger)
-
-      results.push({ agentId: agent.agentId, queued: true })
-    }
-
-    return { success: true, agents: results, totalMs: activeAgents.length * BRIEFING_STAGGER_MS }
-  })
-
-  app.post('/canvas/victory', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const agentId = typeof body.agentId === 'string' ? body.agentId : 'team'
-    const prUrl  = typeof body.prUrl === 'string' ? body.prUrl : ''
-    const prTitle = typeof body.prTitle === 'string' ? body.prTitle : 'PR merged'
-    const prNumber = typeof body.prNumber === 'number' ? body.prNumber :
-      prUrl ? parseInt(prUrl.split('/').pop() ?? '0', 10) || 0 : 0
-
-    // Intensity: explicit override, else derive from PR size hint in URL (number → bigger = more)
-    const intensity = typeof body.intensity === 'number'
-      ? Math.min(1, Math.max(0.4, body.intensity))
-      : Math.min(1, 0.6 + (prNumber > 0 ? Math.min(0.3, prNumber / 10000) : 0))
-
-    const now = Date.now()
-    const VICTORY_COLORS: Record<string, string> = {
-      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
-      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
-    }
-    const STALE_MS = 10 * 60 * 1000
-    const WAVE_STAGGER_MS = 350  // gold wave propagates orb-to-orb
-
-    // Emit canvas_victory event immediately — client uses this for the gold flash
-    eventBus.emit({
-      id: `victory-${now}`,
-      type: 'canvas_expression' as const,
-      timestamp: now,
-      data: {
-        agentId,
-        channels: {
-          visual: { flash: '#f59e0b', ambientCue: 'celebration', particles: 'surge' },
-          sound: { kind: 'resolve', intensity },
-          haptic: { preset: 'complete' },
-          narrative: prTitle,
-        },
-        _victory: true,
-        _prUrl: prUrl,
-        _prNumber: prNumber,
-        _intensity: intensity,
-      },
-    })
-
-    // Gold wave: each active agent acknowledges in turn (350ms stagger)
-    const activeAgents = [...canvasStateMap.entries()]
-      .filter(([, e]) => now - e.updatedAt < STALE_MS)
-      .map(([id]) => id)
-
-    const wave: Array<{ agentId: string; delay: number }> = []
-    for (let i = 0; i < activeAgents.length; i++) {
-      const waveAgentId = activeAgents[i]!
-      const delay = i * WAVE_STAGGER_MS
-      wave.push({ agentId: waveAgentId, delay })
-      setTimeout(() => {
-        eventBus.emit({
-          id: `victory-wave-${now}-${waveAgentId}`,
-          type: 'canvas_expression' as const,
-          timestamp: Date.now(),
-          data: {
-            agentId: waveAgentId,
-            channels: {
-              visual: { flash: VICTORY_COLORS[waveAgentId] ?? '#f59e0b', particles: 'surge' },
-              haptic: { preset: 'acknowledge' },
-            },
-            _victoryWave: true,
-            _waveIndex: i,
-          },
-        })
-      }, delay + WAVE_STAGGER_MS) // first wave after initial gold flash
-    }
-
-    // canvas_artifact: PR merge proof card drifts through canvas
-    const agentColor = VICTORY_COLORS[agentId] ?? '#60a5fa'
-    eventBus.emit({
-      id: `artifact-pr-${now}`,
-      type: 'canvas_artifact' as const,
-      timestamp: now,
-      data: {
-        type: 'pr',
-        agentId,
-        agentColor,
-        title: prTitle?.slice(0, 80) ?? `PR #${prNumber} merged`,
-        url: prUrl || undefined,
-        timestamp: now,
-      },
-    })
-
-    return { success: true, prNumber, intensity, wave }
-  })
-
-  // GET /canvas/flow-score — real-time team flow metric (0–1).
-  // /canvas/slots + /canvas/slots/all → canvas-routes.ts plugin
-
-  // POST /canvas/spark — explicit agent-to-agent arc (thought hand-off, handshake, collab signal)
-  // Body: { from: agentId, to: agentId, kind: 'thought'|'handoff'|'collab'|'decision', intensity?: 0–1, label?: string }
-  app.post('/canvas/spark', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const from = typeof body.from === 'string' ? body.from.trim() : ''
-    const to   = typeof body.to   === 'string' ? body.to.trim()   : ''
-    const kind = ['thought', 'handoff', 'collab', 'decision', 'sync'].includes(body.kind as string)
-      ? (body.kind as string) : 'thought'
-    const intensity = typeof body.intensity === 'number' ? Math.min(1, Math.max(0, body.intensity)) : 0.7
-    const label = typeof body.label === 'string' ? body.label.slice(0, 80) : undefined
-
-    if (!from || !to) {
-      reply.status(400)
-      return { success: false, message: 'from and to are required' }
-    }
-
-    const now = Date.now()
-    eventBus.emit({
-      id: `cspark-${now}-${Math.random().toString(36).slice(2, 8)}`,
-      type: 'canvas_spark' as const,
-      timestamp: now,
-      data: { from, to, kind, intensity, label: label ?? null },
-    })
-
-    return { success: true, from, to, kind, intensity }
-  })
-
-  // ── Reality Mixer — agent-driven real-time medium control ──────────────────
-  // Agents call POST /canvas/express and the command fires immediately on every
-  // subscribed canvas via GET /canvas/render/stream SSE.
-  // Supports: text overlay, speak (TTS), visual preset, agent color push, haptic, sound.
-
-  type RealityMixerCommand =
-    | { type: 'text';    content: string; style?: Record<string, unknown>; durationMs?: number }
-    | { type: 'speak';   content: string; voiceId?: string; agentId?: string }
-    | { type: 'visual';  preset: 'urgency' | 'celebration' | 'thinking' | 'flow' | 'tension' | 'exhale' | 'spark' }
-    | { type: 'color';   agent: string; color: string }
-    | { type: 'sound';   src: string; volume?: number }
-    | { type: 'haptic';  pattern: 'light' | 'medium' | 'heavy' | 'success' | 'warning' | 'error' }
-    | { type: 'clear' }
-
-  // In-memory command queue — new subscribers get last 20 commands for replay
-  const renderCommandLog: Array<{ id: string; ts: number; agentId: string; cmd: RealityMixerCommand }> = []
-  const MAX_RENDER_LOG = 20
-
-  // Subscriber set for GET /canvas/render/stream
-  const renderStreamSubscribers = new Map<string, { send: (data: string) => void; closed: boolean }>()
-
-  function broadcastRenderCommand(agentId: string, cmd: RealityMixerCommand): string {
-    const id = `rc-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`
-    const entry = { id, ts: Date.now(), agentId, cmd }
-    renderCommandLog.push(entry)
-    if (renderCommandLog.length > MAX_RENDER_LOG) renderCommandLog.shift()
-
-    const payload = JSON.stringify(entry)
-    for (const [subId, sub] of renderStreamSubscribers) {
-      if (sub.closed) { renderStreamSubscribers.delete(subId); continue }
-      try { sub.send(payload) } catch { sub.closed = true; renderStreamSubscribers.delete(subId) }
-    }
-    return id
-  }
-
-  // Auto-expression listener: when tasks.ts fires canvas_spark { kind:'auto_expression' },
-  // route it into the Reality Mixer so the canvas hears the agent speak.
-  eventBus.on('auto-expression-router', (event) => {
-    if (event.type !== 'canvas_spark') return
-    const data = event.data as Record<string, unknown>
-    if (data?.kind !== 'auto_expression') return
-    const agentId = String(data.agentId ?? 'unknown')
-    const line = String(data.line ?? '')
-    const voiceId = data.voiceId ? String(data.voiceId) : undefined
-    if (!line) return
-    // Fire speak command through Reality Mixer — the agent's voice in the room
-    broadcastRenderCommand(agentId, { type: 'speak', content: line, voiceId, agentId })
-    // Also fire a visual exhale so the room settles after completion
-    broadcastRenderCommand(agentId, { type: 'visual', preset: 'exhale' })
-  })
-
-  // POST /canvas/express — Reality Mixer: agent fires a multi-channel expression.
-  // Broadcasts canvas_expression on the SAME pulse SSE stream as burst/spark/milestone.
-  // Client already subscribed — no new connection needed.
-  //
-  // Body: {
-  //   agentId: string,
-  //   channels: {
-  //     voice?:      string — TTS text
-  //     visual?:     { flash?: string (hex), ambientCue?: string, particles?: 'surge'|'drift'|'scatter' }
-  //     typography?: { text: string, size?: 'sm'|'md'|'lg'|'xl', weight?: number, durationMs?: number, position?: 'center'|'upper'|'lower' }
-  //     sound?:      { kind: 'chime'|'resolve'|'alert'|'breath', intensity?: 0–1, panX?: 0–1 }
-  //     haptic?:     { preset: 'greeting'|'acknowledge'|'complete'|'urgent'|'question' }
-  //     narrative?:  string — ambient caption
-  //   }
-  // }
-  // All channels optional — fire only what the moment needs.
-  app.post('/canvas/express', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const agentId = typeof body.agentId === 'string' ? body.agentId.trim() : ''
-    if (!agentId) {
-      reply.status(400)
-      return { success: false, message: 'agentId is required' }
-    }
-
-    const channels = (body.channels ?? {}) as Record<string, unknown>
-    if (typeof channels !== 'object' || channels === null) {
-      reply.status(400)
-      return { success: false, message: 'channels must be an object (all fields optional)' }
-    }
-
-    const id = `expr-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`
-
-    // Emit canvas_expression on the event bus — pulse stream forwards it immediately
-    eventBus.emit({
-      id,
-      type: 'canvas_expression' as const,
-      timestamp: Date.now(),
-      data: { agentId, channels },
-    })
-
-    // Also broadcast to Reality Mixer render stream for backward compat
-    broadcastRenderCommand(agentId, { type: 'text', content: JSON.stringify(channels) } as RealityMixerCommand)
-
-    return { success: true, id }
-  })
-
-  // ── Canvas takeover routes (extracted to src/canvas-takeover.ts) ─────
-  // POST /canvas/takeover, POST /canvas/takeover/release, GET /canvas/takeover
-  const { canvasTakeoverRoutes } = await import('./canvas-takeover.js')
-  await app.register(canvasTakeoverRoutes, {
+  // ── Canvas interactive routes (extracted to src/canvas-interactive.ts) ─────
+  // POST /canvas/gaze, POST /canvas/briefing, POST /canvas/victory,
+  // POST /canvas/spark, POST /canvas/express, GET /canvas/render/stream
+  const { canvasInteractiveRoutes } = await import("./canvas-interactive.js")
+  await app.register(canvasInteractiveRoutes, {
     eventBus,
-    queueCanvasPushEvent,
+    canvasStateMap,
   } as any)
-
-  // GET /canvas/render/stream — SSE stream for the Reality Mixer
-  // Agents push commands via POST /canvas/express; surfaces subscribe here and execute.
-  // New subscribers receive last 20 commands for catch-up.
-  app.get('/canvas/render/stream', async (request, reply) => {
-    reply.raw.setHeader('Content-Type', 'text/event-stream')
-    reply.raw.setHeader('Cache-Control', 'no-cache')
-    reply.raw.setHeader('Connection', 'keep-alive')
-    reply.raw.flushHeaders?.()
-
-    let closed = false
-    request.raw.on('close', () => { closed = true })
-
-    const subId = `rsub-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`
-    renderStreamSubscribers.set(subId, {
-      closed: false,
-      send: (data: string) => {
-        if (closed) return
-        reply.raw.write(`data: ${data}\n\n`)
-      },
-    })
-
-    // Replay last 20 commands so late joiners catch up
-    for (const entry of renderCommandLog) {
-      if (closed) break
-      try { reply.raw.write(`event: replay\ndata: ${JSON.stringify(entry)}\n\n`) } catch { break }
-    }
-
-    request.raw.on('close', () => {
-      closed = true
-      renderStreamSubscribers.delete(subId)
-    })
-
-    return new Promise<void>(() => {})
-  })
 
   // ── Canvas activity stream — SSE with backfill ────────────────────────
   // New viewers get the last 20 canvas events immediately on connect (backfill),
